Issue #347: Rename file as per review
