+++ fitbenchmarking/results_processing/plots.py
